途中だけどだいたい動くようになった気がする
